+++ CHANGELOG.md
@@ -8,6 +8,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 
 
+## 3.2.0 (2025-07-07)
+
+### Added
+
+ - support for pitch fitting to avoid particular pitches
+
+ - support for pitch fitting to use a backup strategy if the main one doesn't succeed
+
+
 ## 3.1.0 (2024-02-15)
 
 ### Added